WIP: AWS Lambda job to publish or delete uploads
This is a small rust binary which, when files are uploaded to S3, will
inspect them and either make them publicly accessible, or delete them.

It needs an upload trigger configuring, and should be set up for a
bucket whose default behaviour is public can upload (but not overwrite),
but no one can read. It should be set to run as an account which has
permissions to GetObject, DeleteObject, and SetObjectAcl.

Currently the heuristic for "Is this file good to publish" is "Is it a
zip file with a valid central directory?" - we should improve on that.

The idea would be that the bootstrapper script would upload, and then
block until the file was visible before allowing downstream jobs to be
triggered.

This can be built by running `src/rust/pantsbuild-s3-publisher/build.sh`
and then the produced zip file can be uploaded as a "Use custom runtime
in function code or layer" AWS Lambda.